+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.3.0] - 2026-03-16
+
 ### Changed
 - update AI evaluator guidelines and logic